+++ HISTORY.md
@@ -1,5 +1,13 @@
 # History
 
+## not yet released, version 1.1.2
+
+- Added an event property `event.firstTarget`, always containing the element
+  where a gesture started.
+- Fixed the `hammer.input` event blocking the emit of real events.
+- Fixed `panend` not being emitted when `pan` was used too.
+
+
 ## 2015-01-02, version 1.1.1
 
 - Fixed the module not working via commonjs.